Добавлен cpp файл
Кроме того, изменен тип pin_number на const unsigned char
